+++ archive/org.osgi.test.cases.diagnostics/src/org/osgi/test/cases/diagnostics/tb2/DoorStatus.java
@@ -56,9 +56,9 @@ public class DoorStatus implements Status {
    */
   public byte getStatus() {
     if (errorCode != -1) {
-      return Status.STATUS_FAILED;
+      return (byte) Status.STATUS_FAILED;
     }
-    return Status.STATUS_OK;
+    return (byte) Status.STATUS_OK;
   }
 
   /*
